requirements.txt
# llama-cpp-python
bs4
transformers
faiss-cpu
torch
gradio
requests
numpy
google-generativeai